CR changes: Small refactor.
